Improved CSS Structure & Added Comments
Added in a global CSS file that will be referenced in all pages of the website. This will make it easier to change the style of the overall website, such as changing the sizes of headers and footers, changing colours, etc.

Also, comments were added to improve readability and further make it easier to make changes and implement new pages onto the website.